Refactored parser (backwards compatible):
The Parser is now split into a tokenizer and a decoder.

How it works is this: The input is fed to the tokenizer, which
returns an array of tokens. A token may be an operator, or
an operand. The tokenizer already filters out comments and empty
lines as well as anything else undesirable. If something unknown
appears, the tokenizer logs an error.

Then, the parser reads the first token - which must be an operator.
It then tries to find that operator in the opcodeInfo table. If found,
it fetches the required number of operands from the token list, and
feeds it to the decoder.

The decoder in turn modifies the opcode to reflect the (possible)
indirection of the parameter. It then returns a list containing
the final opcode (including variant) and a decoded list of it's
operands.

The parser then adds those to the memory image, and continues with
the next operator, until it runs out of tokens.

I think that this design, while a bit harder to understand, will be
better in the future when we want to add syntactical extensions, such
as aliases, and macros.