Parser: Implemented jump labels
For this to work, a few changes needed to be done in various parts
of the parser module.

First, the tokenizer needed to know about the new token type. This
was easy, you can just add another regexp with a corresponding callback
to the list in tokenize(). However, since we now have two different
token types coming out of the tokenizer, a type attribute had to be
added.

Then, in the parser, we need to find the jump labels in the token
stream, and remember the address they appeared on. So in a first
loop, we scan for jump labels, and note down their names and addresses.

In the second loop, we just exclude the jump label tokens from
further processing.

The last part that needed adjusting was the decoder. Since we now
can have labels as operands, the decder needs to know about them as
well. So we just pass the labels as a third parameter and let decode()
figure out what to do if an operand is not an integer.